Generacion de clase Tarifa (COMPLETA). Correcion de tipos de datos long por int en clase medidores.
Gabriel Vargas.

+++ InterlocutorComercial/Medidor.cpp
@@ -58,7 +58,7 @@ string Medidor::toStringMedidor()
  
 }
 
-long Medidor::getId() {
+int Medidor::getId() {
     return this->_id;
 }
